New array-growing macros: sgrowarray and sgrowarrayn.
The idea of these is that they centralise the common idiom along the
lines of

   if (logical_array_len >= physical_array_size) {
       physical_array_size = logical_array_len * 5 / 4 + 256;
       array = sresize(array, physical_array_size, ElementType);
   }

which happens at a zillion call sites throughout this code base, with
different random choices of the geometric factor and additive
constant, sometimes forgetting them completely, and generally doing a
lot of repeated work.

The new macro sgrowarray(array,size,n) has the semantics: here are the
array pointer and its physical size for you to modify, now please
ensure that the nth element exists, so I can write into it. And
sgrowarrayn(array,size,n,m) is the same except that it ensures that
the array has size at least n+m (so sgrowarray is just the special
case where m=1).

Now that this is a single centralised implementation that will be used
everywhere, I've also gone to more effort in the implementation, with
careful overflow checks that would have been painful to put at all the
previous call sites.

This commit also switches over every use of sresize(), apart from a
few where I really didn't think it would gain anything. A consequence
of that is that a lot of array-size variables have to have their types
changed to size_t, because the macros require that (they address-take
the size to pass to the underlying function).

+++ dialog.h
@@ -422,8 +422,8 @@ struct controlset {
     char *boxname;		       /* internal short name of controlset */
     char *boxtitle;		       /* title of container box */
     int ncolumns;		       /* current no. of columns at bottom */
-    int ncontrols;		       /* number of `union control' in array */
-    int ctrlsize;		       /* allocated size of array */
+    size_t ncontrols;                  /* number of `union control' in array */
+    size_t ctrlsize;                   /* allocated size of array */
     union control **ctrls;	       /* actual array */
 };
 
@@ -434,11 +434,11 @@ typedef void (*ctrl_freefn_t)(void *);    /* used by ctrl_alloc_with_free */
  * controls.
  */
 struct controlbox {
-    int nctrlsets;		       /* number of ctrlsets */
-    int ctrlsetsize;		       /* ctrlset size */
+    size_t nctrlsets;		       /* number of ctrlsets */
+    size_t ctrlsetsize;		       /* ctrlset size */
     struct controlset **ctrlsets;      /* actual array of ctrlsets */
-    int nfrees;
-    int freesize;
+    size_t nfrees;
+    size_t freesize;
     void **frees;		       /* array of aux data areas to free */
     ctrl_freefn_t *freefuncs;          /* parallel array of free functions */
 };
