+++ terminal.h
@@ -270,7 +270,7 @@ struct terminal_tag {
     bidi_char *wcFrom, *wcTo;
     int wcFromTo_size;
     struct bidi_cache_entry *pre_bidi_cache, *post_bidi_cache;
-    int bidi_cache_size;
+    size_t bidi_cache_size;
 
     /*
      * We copy a bunch of stuff out of the Conf structure into local
